infra: move Neo unit tests to a new file
